[FancyZones] Clean up (#11893)
* moved window messages
* moved util
* moved monitor utils
