+++ src/modules/fancyzones/FancyZonesLib/FancyZonesLib.vcxproj.filters
@@ -81,6 +81,9 @@
     <ClInclude Include="CallTracer.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="MonitorUtils.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="pch.cpp">
@@ -138,6 +141,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="CallTracer.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="MonitorUtils.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
